+++ pkg/db/gorm.go
@@ -94,9 +94,10 @@ func Close(conn *gorm.DB) {
 
 // connect reads environment variables for DB configuration and attempts to open the connection
 func connect(opts *ConnectionOptions) (*gorm.DB, error) {
-	connectString := fmt.Sprintf("host=%s port=%s dbname=%s user=%s password=%s sslmode=%s",
-		opts.Host, opts.Port, opts.DatabaseName, opts.User, opts.Password, opts.SSLMode)
-	logging.LogDebugf("Attempting to connect to DB: host = %s, port = %s, dbname = %s", opts.Host, opts.Port, opts.DatabaseName)
+	connectString := fmt.Sprintf("host=%s port=%s dbname=%s user=%s password=%s sslmode=%s sslrootcert=%s",
+		opts.Host, opts.Port, opts.DatabaseName, opts.User, opts.Password, opts.SSLMode, opts.SSLRootCertPath)
+	logging.LogDebugf("Attempting to connect to DB: host = %s, port = %s, dbname = %s, ssl-mode = %s, sslrootcert = %s",
+		opts.Host, opts.Port, opts.DatabaseName, opts.SSLMode, opts.SSLRootCertPath)
 	return gorm.Open("postgres", connectString)
 }